- Issue #79 (missing folder test)

+++ src/jarvis/jarvis.py
@@ -114,13 +114,14 @@ def clean_folders():
     @return None
     """
     for folder in ["./diagrams", "./log"]:
-        for filename in os.listdir(folder):
-            file_path = os.path.join(folder, filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.remove(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as ex:
-                # function called before Logger initialization.
-                print('[ERROR] Failed to delete %s. Reason: %s' % (file_path, ex))
+        if os.path.isdir(folder):
+            for filename in os.listdir(folder):
+                file_path = os.path.join(folder, filename)
+                try:
+                    if os.path.isfile(file_path) or os.path.islink(file_path):
+                        os.remove(file_path)
+                    elif os.path.isdir(file_path):
+                        shutil.rmtree(file_path)
+                except Exception as ex:
+                    # function called before Logger initialization.
+                    print('[ERROR] Failed to delete %s. Reason: %s' % (file_path, ex))
